-transload.me: MOCH fixed account type parse bug api returns double not integer RE tester blacksun refs #76501
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@34462 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/plugins/hoster/TransloadMe.java
@@ -218,7 +218,7 @@ public class TransloadMe extends PluginForHost {
         /*
          * Free users = They have no package (==no balance left) --> Accept them but set zero traffic left.
          */
-        if (balance == null || Integer.parseInt(balance) <= 0) {
+        if (balance == null || Double.parseDouble(balance) <= 0) {
             account.setType(AccountType.FREE);
             ai.setStatus("Registered (free) account");
             /* Free accounts have no traffic - set this so they will not be used (accidently) but still accept them. */
